// client/components/AddDescription/AddDescription.styled.tsx
import styled from "styled-components";

export const AddDescription = {
  AddDescription: styled.div``,
  Header: styled.header`
    padding-top: 32px;
    padding-bottom: 32px;
  `,
  XContainer: styled.div`
    padding: 0 24px;
  `,
  Content: styled.div`
    padding-top: 32px;
    padding-bottom: 32px;
  `,
  Line: styled.div`
    height: 1px;
    background-color: #e2e0dc;
  `,
  Buttons: styled.div`
    margin-top: 32px;
    display: flex;
    justify-content: space-between;
    align-items: center;
    gap: 24px;
  `,
  Editor: styled.div`
    display: grid;
    height: 200px;
  `,
};